Split UserController methods from DevNestController
